requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM & AI
google-generativeai==0.3.2
groq==0.4.1

# Data processing
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2

# Visualization & Reports
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
reportlab==4.0.9

# Frontend
streamlit==1.30.0

# Database
sqlalchemy==2.0.25
# sqlite3 is built-in to Python

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
tqdm==4.66.1

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
